knowledge/requirements.txt
pinecone==5.4.2
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
